+++ src/main/java/net/mcreator/scpadditions/config/ScpAdditionsModulesConfig.java
@@ -141,7 +141,10 @@ public final class ScpAdditionsModulesConfig {
 
 	public static final class Audio {
 		@SerializedName("enter_sound_enabled")
-		public boolean enterSoundEnabled = true;
+        public boolean enterSoundEnabled = true;
+
+		@SerializedName("save_game_sound_enabled")
+		public boolean saveGameSoundEnabled = true;
 
 		@SerializedName("replace_player_hurt_sounds")
 		public boolean replacePlayerHurtSounds = true;